Casual benchmarking, done by repeatedly multiplying two 100x100 matrices,
suggests that throughput increases more than 100-fold.

Before

```
$ ts-node matmul_bench.ts
document is not defined
time: 2.625s  count: 4  throughput: 3.8095238095238093/s
time: 1.0309998989105225s  count: 4  throughput: 3.8797287994178054/s
time: 1.0380001068115234s  count: 4  throughput: 3.8535641506695013/s
time: 1.0469999313354492s  count: 4  throughput: 3.8204396010780983/s
time: 1.0780000686645508s  count: 4  throughput: 3.710574902796884/s
time: 1.062000036239624s  count: 4  throughput: 3.7664782142224533/s
```

After:


```
$ ts-node matmul_bench.ts
document is not defined
time: 0.046000003814697266s  count: 217  throughput: 217.39128631995771/s
time: 0.5s  count: 416  throughput: 434/s
time: 0.9539999961853027s  count: 429  throughput: 436.05870195328293/s
time: 0.9840002059936523s  count: 432  throughput: 435.9755184876124/s
time: 0.9849998950958252s  count: 434  throughput: 438.5787269124258/s
time: 1.0150001049041748s  count: 432  throughput: 427.5861627038684/s
```

Benchmark script: https://gist.github.com/piscisaureus/4afe3a5d2a9b97e57835a8c4d826243c

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/tensorflow/tfjs-core/961)
<!-- Reviewable:end -->