@dsmilkov 2018-04-17
<img class="emoji" title=":lgtm_strong:" alt=":lgtm_strong:" align="absmiddle" src="https://reviewable.io/lgtm_strong.png" height="20" width="61"/> Thank you x 100! That TODO was long overdue. The reason we haven't optimized the CPU backend this long is because we are mostly using it as a test against our WebGL backend, though I agree it's no a good excuse. Thanks again!

---

Review status: 0 of 1 files reviewed at latest revision, all discussions resolved, some commit checks failed.

---



*Comments from [Reviewable](https://beta.reviewable.io/reviews/tensorflow/tfjs-core/961#-:-LAG44JfEqXwWe6bXd3g:bq5w0rj)*
<!-- Sent from Reviewable.io -->