@piscisaureus 2018-04-17
> The reason we haven't optimized the CPU backend this long is because we are mostly using it as a test against our WebGL backend, though I admit it's not a good excuse. Thanks again!

That's okay, don't apologize. It just became noticeable for use because in node we have to use the CPU backend.

> Ignore the googlebot comment since I synched your branch with master.

I rebased, so the CLA checker is  happy now.